notificationservice/requirements.txt
django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pika==1.3.2
twilio==8.10.2